chore: add 'WriteEvery' to core pprof, remove duplicated code
