+++ libs/pprof/pprof.go
@@ -15,6 +15,7 @@ package pprof
 import (
 	"fmt"
 	"net/http"
+	"time"
 
 	// import pprof globally because it's used to init the package
 	// and this comment is mostly here as well in order to make
@@ -24,7 +25,6 @@ import (
 	"path/filepath"
 	"runtime"
 	"runtime/pprof"
-	"time"
 
 	"code.vegaprotocol.io/vega/libs/config/encoding"
 	vgfs "code.vegaprotocol.io/vega/libs/fs"
@@ -53,6 +53,8 @@ type Config struct {
 	// To just read the current rate, pass rate < 0.
 	// (For n>1 the details of sampling may change.)
 	MutexProfileFraction int `long:"mutex-profile-fraction"`
+	// Write the profiles to disk every WriteEvery interval
+	WriteEvery encoding.Duration `long:"write-every"  description:"write pprof files at this interval; if 0 only write on shutdown"`
 }
 
 // Pprofhandler is handling pprof profile management.
@@ -60,6 +62,8 @@ type Pprofhandler struct {
 	Config
 
 	log            *logging.Logger
+	stop           chan struct{}
+	done           chan struct{}
 	memprofilePath string
 	cpuprofilePath string
 }
@@ -73,6 +77,7 @@ func NewDefaultConfig() Config {
 		ProfilesDir:          "/tmp",
 		BlockProfileRate:     0,
 		MutexProfileFraction: 0,
+		WriteEvery:           encoding.Duration{Duration: 15 * time.Minute},
 	}
 }
 
@@ -85,30 +90,65 @@ func New(log *logging.Logger, config Config) (*Pprofhandler, error) {
 	runtime.SetBlockProfileRate(config.BlockProfileRate)
 	runtime.SetMutexProfileFraction(config.MutexProfileFraction)
 
-	t := time.Now()
-	memprofileFile := fmt.Sprintf("%s-%s%s", memprofileName, t.Format("2006-01-02-15-04-05"), profileExt)
-	cpuprofileFile := fmt.Sprintf("%s-%s%s", cpuprofileName, t.Format("2006-01-02-15-04-05"), profileExt)
-
 	p := &Pprofhandler{
-		log:            log,
-		Config:         config,
-		memprofilePath: filepath.Join(config.ProfilesDir, pprofDir, memprofileFile),
-		cpuprofilePath: filepath.Join(config.ProfilesDir, pprofDir, cpuprofileFile),
+		log:    log,
+		Config: config,
+		stop:   make(chan struct{}),
+		done:   make(chan struct{}),
 	}
 
 	// start the pprof http server
 	go func() {
-		p.log.Error("pprof web server closed", logging.Error(http.ListenAndServe("localhost:6060", nil)))
+		p.log.Error("pprof web server closed", logging.Error(http.ListenAndServe(fmt.Sprintf("localhost:%d", config.Port), nil)))
 	}()
 
-	// start cpu and mem profilers
-	if err := vgfs.EnsureDir(filepath.Join(config.ProfilesDir, pprofDir)); err != nil {
-		p.log.Error("Could not create CPU profile file",
-			logging.String("path", p.cpuprofilePath),
+	// make sure profile dir exists
+	profDir := filepath.Join(config.ProfilesDir, pprofDir)
+	if err := vgfs.EnsureDir(profDir); err != nil {
+		p.log.Error("Could not create profile dir",
+			logging.String("path", profDir),
 			logging.Error(err),
 		)
 		return nil, err
 	}
+
+	go p.runProfiling()
+
+	return p, nil
+}
+
+func (p *Pprofhandler) runProfiling() error {
+	defer close(p.done)
+	// If WriteEvery is 0, make a ticker that never ticks
+	tick := make(<-chan time.Time)
+	if p.WriteEvery.Duration > 0 {
+		tick = time.NewTicker(p.WriteEvery.Duration).C
+	}
+
+	for {
+		if err := p.startProfiling(); err != nil {
+			return err
+		}
+
+		select {
+		case <-p.stop:
+			return p.stopProfiling()
+		case <-tick:
+		}
+
+		if err := p.stopProfiling(); err != nil {
+			return err
+		}
+	}
+}
+
+func (p *Pprofhandler) startProfiling() error {
+	t := time.Now()
+	memprofileFile := fmt.Sprintf("%s-%s%s", memprofileName, t.Format("2006-01-02-15-04-05"), profileExt)
+	cpuprofileFile := fmt.Sprintf("%s-%s%s", cpuprofileName, t.Format("2006-01-02-15-04-05"), profileExt)
+
+	p.memprofilePath = filepath.Join(p.ProfilesDir, pprofDir, memprofileFile)
+	p.cpuprofilePath = filepath.Join(p.ProfilesDir, pprofDir, cpuprofileFile)
 
 	profileFile, err := os.Create(p.cpuprofilePath)
 	if err != nil {
@@ -116,11 +156,11 @@ func New(log *logging.Logger, config Config) (*Pprofhandler, error) {
 			logging.String("path", p.cpuprofilePath),
 			logging.Error(err),
 		)
-		return nil, err
+		return err
 	}
 	pprof.StartCPUProfile(profileFile)
 
-	return p, nil
+	return nil
 }
 
 // ReloadConf update the configuration of the pprof package.
@@ -142,6 +182,12 @@ func (p *Pprofhandler) ReloadConf(cfg Config) {
 
 // Stop is meant to be use to stop the pprof profile, should be use with defer probably.
 func (p *Pprofhandler) Stop() error {
+	close(p.stop)
+	<-p.done
+	return nil
+}
+
+func (p *Pprofhandler) stopProfiling() error {
 	// stop cpu profile once the memory profile is written
 	defer pprof.StopCPUProfile()
